\name{WIdataRipper-package}
\alias{WIdataRipper-package}
\alias{WIdataRipper}
\docType{package}
\title{
WaterInfo data ripper.
}
\description{
This package is intended to perform basic data and information retrieval functions for the NSW government stream and rivers network database. The functions are limited for ease of use. Please see the source code for further information about the available methods for data retrieval.
}
\details{
\tabular{ll}{
Package: \tab WIdataRipper\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2011-08-24\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Jason Lessels

Maintainer: Jason Lessels <jlessels@gmail.com>

}

\keyword{ package }
\examples{
getSiteVariables(410044)
}
